backend/requirements.txt
flask
flask-cors
numpy
pandas
librosa
soundfile
scikit-learn
praat-parselmouth
